[release-v0.0] Define behavior of retention policy in multi-namespace plans. (#56)
* Define behavior of retention policy in multi-namespace plans. Before it was undefined what would happen if the user provided both the retention policy on the multi-namespace plan and on the namespace plan. Now the namespace plan inherits from the multi-namespace plan if not set.

CEL validation is still used to validate the
retention policy. In particular it should prevent
the user from removing the retention policy once
it is set.


* Added unit tests for retention policy


* Fix linter


---------

+++ internal/controller/testutils/test-setup.go
@@ -47,10 +47,16 @@ func CreateMigPlanNamespace(ctx context.Context, client client.Client, namespace
 }
 
 func DeleteKubeVirtNamespace(ctx context.Context, client client.Client, namespace string) {
+	if client == nil {
+		return
+	}
 	Expect(client.Delete(ctx, &corev1.Namespace{ObjectMeta: metav1.ObjectMeta{Name: namespace}})).To(Succeed())
 }
 
 func DeleteMigPlanNamespace(ctx context.Context, client client.Client, namespace string) {
+	if client == nil {
+		return
+	}
 	Expect(client.Delete(ctx, &corev1.Namespace{ObjectMeta: metav1.ObjectMeta{Name: namespace}})).To(Succeed())
 }
